use 64bit int and float

+++ entities/event.go
@@ -1,7 +1,7 @@
 package entities
 
 type Event struct {
-	ID    uint   `gorm:"primaryKey" json:"id"`
+	ID    uint64 `gorm:"primaryKey" json:"id"`
 	Title string `json:"title"`
 	Start string `json:"start"`
 	End   string `json:"end"`
